+++ kitty/options/to-c-generated.h
@@ -136,6 +136,19 @@ convert_from_opts_cursor_stop_blinking_after(PyObject *py_opts, Options *opts) {
 }
 
 static void
+convert_from_python_scrollback_indicator_opacity(PyObject *val, Options *opts) {
+    opts->scrollback_indicator_opacity = PyFloat_AsFloat(val);
+}
+
+static void
+convert_from_opts_scrollback_indicator_opacity(PyObject *py_opts, Options *opts) {
+    PyObject *ret = PyObject_GetAttrString(py_opts, "scrollback_indicator_opacity");
+    if (ret == NULL) return;
+    convert_from_python_scrollback_indicator_opacity(ret, opts);
+    Py_DECREF(ret);
+}
+
+static void
 convert_from_python_scrollback_pager_history_size(PyObject *val, Options *opts) {
     opts->scrollback_pager_history_size = PyLong_AsUnsignedLong(val);
 }
@@ -1131,6 +1144,8 @@ convert_opts_from_python_opts(PyObject *py_opts, Options *opts) {
     convert_from_opts_cursor_blink_interval(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_cursor_stop_blinking_after(py_opts, opts);
+    if (PyErr_Occurred()) return false;
+    convert_from_opts_scrollback_indicator_opacity(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_scrollback_pager_history_size(py_opts, opts);
     if (PyErr_Occurred()) return false;