+++ kitty/state.h
@@ -53,7 +53,7 @@ typedef struct {
     float macos_thicken_font;
     WindowTitleIn macos_show_window_title_in;
     char *bell_path, *bell_theme;
-    float background_opacity, dim_opacity;
+    float background_opacity, dim_opacity, scrollback_indicator_opacity;
     float text_contrast, text_gamma_adjustment;
     bool text_old_gamma;